test reorganized so more understandable now
