+++ tests/test_detection_match.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 from astropy.table import Table
 import btk
 import btk.metrics
